backend/requirements.txt
# Django REST Framework Backend
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
Pillow==10.1.0
python-decouple==3.8
requests==2.31.0

# Auth (JWT)
djangorestframework-simplejwt==5.3.1
